back/requirements.txt
Flask==2.2.2
pyjwt==2.4.0
werkzeug==2.2.2